+++ src/metadata-builder/EntityMetadataValidator.ts
@@ -8,6 +8,8 @@ import {ColumnType} from "../driver/types/ColumnTypes";
 import {MongoDriver} from "../driver/mongodb/MongoDriver";
 import {SqlServerDriver} from "../driver/sqlserver/SqlServerDriver";
 import {MysqlDriver} from "../driver/mysql/MysqlDriver";
+import {NoConnectionOptionError} from "../error/NoConnectionOptionError";
+import {InitializedRelationError} from "../error/InitializedRelationError";
 
 /// todo: add check if there are multiple tables with the same name
 /// todo: add checks when generated column / table names are too long for the specific driver
@@ -88,10 +90,13 @@ export class EntityMetadataValidator {
                 throw new Error(`Error in ${entityMetadata.name} entity. There can be only one auto-increment column in MySql table.`);
         }*/
 
+        // for mysql we are able to not define a default selected database, instead all entities can have their database
+        // defined in their decorators. To make everything work either all entities must have database define and we
+        // can live without database set in the connection options, either database in the connection options must be set
         if (driver instanceof MysqlDriver) {
             const metadatasWithDatabase = allEntityMetadatas.filter(metadata => metadata.database);
             if (metadatasWithDatabase.length === 0 && !driver.database)
-                throw new Error(`Database not specified`);
+                throw new NoConnectionOptionError("database");
         }
 
         if (driver instanceof SqlServerDriver) {
@@ -99,6 +104,22 @@ export class EntityMetadataValidator {
             if (charsetColumns.length > 1)
                 throw new Error(`Character set specifying is not supported in Sql Server`);
         }
+
+        // check if relations are all without initialized properties
+        const entityInstance = entityMetadata.create();
+        entityMetadata.relations.forEach(relation => {
+            if (relation.isManyToMany || relation.isOneToMany) {
+
+                // we skip relations for which persistence is disabled since initialization in them cannot harm somehow
+                if (relation.persistenceEnabled === false)
+                    return;
+
+                // get entity relation value and check if its an array
+                const relationInitializedValue = relation.getEntityValue(entityInstance);
+                if (relationInitializedValue instanceof Array)
+                    throw new InitializedRelationError(relation);
+            }
+        });
 
         // validate relations
         entityMetadata.relations.forEach(relation => {