+++ src/schema-builder/RdbmsSchemaBuilder.ts
@@ -345,9 +345,15 @@ export class RdbmsSchemaBuilder implements SchemaBuilder {
             if (!table)
                 return;
 
-            const metadataPrimaryColumns = metadata.columns.filter(column => column.isPrimary && !column.isGenerated);
-            const addedPrimaryColumns = metadataPrimaryColumns.filter(metadataPrimaryColumn => {
-                return !!table.primaryKeyWithoutGenerated && !table.primaryKeyWithoutGenerated.columnNames.find(columnName => columnName === metadataPrimaryColumn.databaseName);
+            const metadataPrimaryColumns = metadata.columns.filter(column => column.isPrimary);
+            const addedKeys = metadataPrimaryColumns
+                .filter(primaryKey => {
+                    return !table.primaryKeys.find(dbPrimaryKey => dbPrimaryKey.columnName === primaryKey.databaseName);
+                })
+                .map(primaryKey => new TablePrimaryKey("", primaryKey.databaseName));
+
+            const droppedKeys = table.primaryKeys.filter(primaryKeySchema => {
+                return !metadataPrimaryColumns.find(primaryKeyMetadata => primaryKeyMetadata.databaseName === primaryKeySchema.columnName);
             });
             const isPrimaryKeyDropped = !metadataPrimaryColumns.find(metadataPrimaryColumn => {
                 return !!table.primaryKeyWithoutGenerated && !!table.primaryKeyWithoutGenerated.columnNames.find(columnName => columnName === metadataPrimaryColumn.databaseName);