+++ src/schema-builder/table/Table.ts
@@ -9,6 +9,9 @@ import {EntityMetadata} from "../../metadata/EntityMetadata";
 import {TableUtils} from "../util/TableUtils";
 import {TableUnique} from "./TableUnique";
 import {TableCheck} from "./TableCheck";
+import {AbstractSqliteDriver} from "../../driver/sqlite-abstract/AbstractSqliteDriver";
+import {MysqlDriver} from "../../driver/mysql/MysqlDriver";
+import {OracleDriver} from "../../driver/oracle/OracleDriver";
 
 /**
  * Table in the database represented in this class.
@@ -220,6 +223,7 @@ export class Table {
             // console.log(tableColumn.isUnique, "!==", columnMetadata.isUnique);
             // console.log(tableColumn.isGenerated, "!==", columnMetadata.isGenerated);
 
+            const skipGenerationCheck = columnMetadata.generationStrategy === "uuid" && (driver instanceof AbstractSqliteDriver || driver instanceof MysqlDriver || driver instanceof OracleDriver);
             return  tableColumn.name !== columnMetadata.databaseName ||
                     tableColumn.type !== driver.normalizeType(columnMetadata) ||
                     // tableColumn.comment !== columnMetadata.comment || // todo
@@ -227,7 +231,7 @@ export class Table {
                     tableColumn.isNullable !== columnMetadata.isNullable ||
                     tableColumn.isUnique !== driver.normalizeIsUnique(columnMetadata) ||
                     // tableColumn.isPrimary !== columnMetadata.isPrimary ||
-                    tableColumn.isGenerated !== columnMetadata.isGenerated ||
+                    (skipGenerationCheck === false && tableColumn.isGenerated !== columnMetadata.isGenerated) ||
                     !this.compareColumnLengths(driver, tableColumn, columnMetadata);
         });
     }